removed warnings for type comparisions

+++ src/basecall.cpp
@@ -12,7 +12,7 @@
 #include "error.h"
 
 void basecall_chunks(std::vector<torch::Tensor> &tensors, std::vector<Chunk *> &chunks, int chunk_size, int batch_size, ModelRunnerBase &model_runner, timestamps_t *ts) {
-    for (int i = 0; i < tensors.size(); ++i) {
+    for (size_t i = 0; i < tensors.size(); ++i) {
         ts->time_accept -= realtime();
         model_runner.accept_chunk(i, tensors[i]);
         ts->time_accept += realtime();
@@ -28,7 +28,7 @@ void basecall_chunks(std::vector<torch::Tensor> &tensors, std::vector<Chunk *> &
     std::vector<DecodedChunk> decoded_chunks = model_runner.decode_chunks(scores, chunks.size());
     ts->time_decode += realtime();
 
-    for (int i = 0; i < chunks.size(); ++i) {
+    for (size_t i = 0; i < chunks.size(); ++i) {
         chunks[i]->seq = decoded_chunks[i].sequence;
         chunks[i]->qstring = decoded_chunks[i].qstring;
         chunks[i]->moves = decoded_chunks[i].moves;
